# backend/docs/swagger.yaml
definitions:
  models.IoTStats:
    properties:
      active_devices:
        type: integer
      error_count:
        type: integer
      rps:
        type: number
      total_events:
        type: integer
    type: object
  models.Issue:
    properties:
      assignee:
        type: string
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      id:
        type: string
      priority:
        description: Low, Medium, High
        type: string
      project_id:
        type: string
      status:
        description: Open, InProgress, Resolved, Closed
        type: string
      x:
        type: number
      "y":
        type: number
      z:
        type: number
    type: object
  models.Organization:
    properties:
      created_at:
        type: string
      google_api_key:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.Project:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
    type: object
  models.Splat:
    properties:
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
  models.SystemLog:
    properties:
      id:
        type: string
      level:
        type: string
      message:
        type: string
      source:
        type: string
      timestamp:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: string
      org_id:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.UserRole:
    enum:
    - owner
    - pm
    - architect
    - foreman
    - specialist
    - worker
    - client
    type: string
    x-enum-varnames:
    - RoleOwner
    - RolePM
    - RoleArchitect
    - RoleForeman
    - RoleSpecialist
    - RoleWorker
    - RoleClient
info:
  contact: {}
paths:
  /issues/{id}/status:
    patch:
      description: Quick status change for an issue
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      - description: New Status
        in: query
        name: status
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Update issue status
      tags:
      - Issues
  /organizations:
    get:
      description: Get list of all organizations (Superadmin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organization'
            type: array
      summary: Get all organizations
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Register a new company/firm
      parameters:
      - description: Organization object
        in: body
        name: org
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organization'
      summary: Create organization
      tags:
      - Admin
  /organizations/{orgID}/users:
    get:
      description: Get list of users for a specific firm
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get users by organization
      tags:
      - Admin
  /projects:
    get:
      description: Get list of all projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
      summary: Get all projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new construction project
      parameters:
      - description: Project object
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
      summary: Create a new project
      tags:
      - Projects
  /projects/{projectID}/issues:
    get:
      description: Get all spatial issues for a project
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Issue'
            type: array
      summary: Get project issues
      tags:
      - Issues
    post:
      consumes:
      - application/json
      description: Mark a defect or issue in 3D space
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Issue object
        in: body
        name: issue
        required: true
        schema:
          $ref: '#/definitions/models.Issue'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Issue'
      summary: Create a spatial issue
      tags:
      - Issues
  /projects/{projectID}/splats:
    post:
      consumes:
      - multipart/form-data
      description: Upload a 3D scan for a project. Returns 202 if accepted for background
        processing.
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Point cloud file
        in: formData
        name: file
        required: true
        type: file
      - description: Splat name
        in: formData
        name: name
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.Splat'
      summary: Upload point cloud (.ply)
      tags:
      - 3D
  /saas/iot/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IoTStats'
      summary: Get IoT Stats (Superadmin)
      tags:
      - SaaS
  /saas/logs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SystemLog'
            type: array
      summary: Get System Logs (Superadmin)
      tags:
      - SaaS
  /users:
    post:
      consumes:
      - application/json
      description: Add a user to an organization
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Create user
      tags:
      - Admin
  /walls/{wallID}:
    get:
      description: Get details of a specific wall including layers
      parameters:
      - description: Wall ID
        in: path
        name: wallID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Get wall details
      tags:
      - Walls
swagger: "2.0"